edited menu, product modify functionality

+++ database/migrations/2021_11_27_070457_create_baskets_table.php
@@ -15,14 +15,14 @@ class CreateBasketsTable extends Migration
     {
         Schema::create('baskets', function (Blueprint $table) {
             $table->id();
-            $table->integer('product_id')->nullable();
-            $table->integer('category_id')->nullable();
+            $table->unsignedBigInteger('product_id')->nullable();
+            $table->unsignedBigInteger('category_id')->nullable();
             $table->tinyInteger('product_type')->nullable();
             $table->double('amount')->nullable();
             $table->string('phone')->nullable();
             $table->string('name')->nullable();
             $table->text('address')->nullable();
-            $table->integer('filial_id')->nullable();
+            $table->unsignedBigInteger('filial_id')->nullable();
             $table->foreignId('bot_user_id')->nullable()->references('chat_id')->on('bot_users')->onDelete('CASCADE');
             $table->boolean('is_served')->default(false);
             $table->boolean('is_finished')->default(false);
